services/worker/app/tasks/media.py
"""
媒体处理任务

处理图片、音频、视频等媒体文件
"""

from typing import Any
from celery import shared_task
import structlog

logger = structlog.get_logger(__name__)


@shared_task(bind=True, max_retries=3)
def process_image(
    self,
    asset_id: str,
    storage_path: str,
    operations: list[str] | None = None,
) -> dict[str, Any]:
    """
    处理图片

    Args:
        asset_id: 资源 ID
        storage_path: 存储路径
        operations: 处理操作列表（thumbnail, resize, compress）

    Returns:
        处理结果
    """
    logger.info("process_image_start", asset_id=asset_id)

    operations = operations or ["thumbnail"]
    results = {}

    try:
        from PIL import Image
        import io

        # TODO: 从对象存储下载图片
        # image_data = download_from_s3(storage_path)
        # image = Image.open(io.BytesIO(image_data))

        for op in operations:
            if op == "thumbnail":
                # 生成缩略图
                # thumb = image.copy()
                # thumb.thumbnail((200, 200))
                # results["thumbnail"] = upload_to_s3(thumb, f"{asset_id}_thumb.jpg")
                results["thumbnail"] = f"{asset_id}_thumb.jpg"

            elif op == "resize":
                # 调整尺寸
                results["resized"] = f"{asset_id}_resized.jpg"

            elif op == "compress":
                # 压缩
                results["compressed"] = f"{asset_id}_compressed.jpg"

        logger.info("process_image_complete", asset_id=asset_id, results=results)

        return {
            "status": "success",
            "asset_id": asset_id,
            "variants": results,
        }

    except Exception as e:
        logger.error("process_image_error", asset_id=asset_id, error=str(e))
        raise self.retry(exc=e, countdown=60)


@shared_task(bind=True, max_retries=3)
def process_audio(
    self,
    asset_id: str,
    storage_path: str,
) -> dict[str, Any]:
    """
    处理音频文件

    Args:
        asset_id: 资源 ID
        storage_path: 存储路径

    Returns:
        处理结果，包含时长、格式等元数据
    """
    logger.info("process_audio_start", asset_id=asset_id)

    try:
        # TODO: 实现音频处理逻辑
        # - 获取时长
        # - 转换格式
        # - 生成波形图

        return {
            "status": "success",
            "asset_id": asset_id,
            "metadata": {
                "duration": 0,
                "format": "mp3",
                "sample_rate": 44100,
            },
        }

    except Exception as e:
        logger.error("process_audio_error", asset_id=asset_id, error=str(e))
        raise self.retry(exc=e, countdown=60)
